pyproject.toml
[project]
name = "hosp-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "emoji<2",
    "fastapi>=0.135.3",
    "ntplib>=0.4.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "pytest>=8.4.2",
    "recognizers-text-suite>=1.0.2a2",
    "requests>=2.32.4",
    "streamlit>=1.56.0",
    "uvicorn>=0.34.2",
]

[project.scripts]
app = "main:app"
